pyproject.toml
[tool.poetry]
name = "olx_rent"
version = "0.1.0"
description = "Monitor new room/appartment offers on OLX"
authors = ["trymzet <michalmzawadzki@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.23.0"
python-dotenv = "^0.12.0"
beautifulsoup4 = "^4.8.2"
pendulum = "^2.1.0"
python-box = "^4.2.2"
pandas = "^1.0.3"
dask = "^2.14.0"
distributed = "^2.14.0"
prefect = "^0.10.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.7.9"
pytest = "^5.4.1"
mypy = "^0.770"
tox = "^3.14.6"
jupyterlab = "^2.0.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
